Changes the comment at the begining of the main file for documentation
Meanwhile changes ware made to R2U2 lib so it would work on ESP32:

### Padding
Paddings ware added to structures to combat differences in memory alignment. These structures are:
- 2u2_mltl_instruction_t
- 2u2_boxq_t
- 2u2_instruction_t
Beyond adding the paddings in the struct definition it is also necessary to add it when the structrue is created with {**values}
Tests ware only made in past time formulas, some other structures may also need padding when different formulas are used

### Memory bound sizes
Standard monitor implementation is to big for the internal memory of most microcontrollers, so alternatives sizes ware added to implement the monitor and are know used in the library code in place of the old larger values.
These values ware placed as defines in internals/bounds.h